Move buyersguide home page to separate file
This "products" file was getting long and just did not fit.
