+++ network-api/networkapi/wagtailpages/pagemodels/buyersguide/products.py
@@ -2,34 +2,25 @@ import re
 import json
 
 from bs4 import BeautifulSoup
-from datetime import datetime
-
 from django.apps import apps
 from django.conf import settings
-from django.core.cache import cache
 from django.core.exceptions import ValidationError
 from django.core.validators import int_list_validator
 from django.db import Error, models
 from django.db.models import F
 from django.http import HttpResponse, HttpResponseNotAllowed, HttpResponseServerError, HttpResponseNotFound
-from django.shortcuts import get_object_or_404, redirect, render
 from django.templatetags.static import static
 from django.utils import timezone
 from django.utils.text import slugify
-from django.utils.translation import gettext, pgettext
-
+from django.utils.translation import gettext
 from modelcluster.fields import ParentalKey
-
 from wagtail.admin.edit_handlers import InlinePanel, FieldPanel, MultiFieldPanel, PageChooserPanel
-from wagtail.contrib.routable_page.models import RoutablePageMixin, route
-from wagtail.core.models import Locale, Orderable, Page, TranslatableMixin
+from wagtail.core.models import Orderable, Page, TranslatableMixin
 from wagtail.core.fields import RichTextField
-
 from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.search import index
 from wagtail.snippets.edit_handlers import SnippetChooserPanel
 from wagtail.snippets.models import register_snippet
-
 from wagtail_localize.fields import SynchronizedField, TranslatableField
 from wagtail_airtable.mixins import AirtableMixin
 
@@ -38,18 +29,17 @@ from networkapi.wagtailpages.fields import ExtendedYesNoField
 from networkapi.wagtailpages.pagemodels.mixin.foundation_metadata import (
     FoundationMetadataPageMixin
 )
-from networkapi.wagtailpages.pagemodels.customblocks.base_rich_text_options import base_rich_text_options
-from networkapi.wagtailpages.templatetags.localization import relocalize_url
-from networkapi.wagtailpages.utils import (
-    insert_panels_after,
-    get_default_locale,
-    get_locale_from_request,
-    get_original_by_slug,
-    TitleWidget
+from networkapi.wagtailpages.pagemodels.buyersguide.utils import (
+    get_categories_for_locale,
 )
-# TODO: Move this util function
+from networkapi.wagtailpages.pagemodels.customblocks.base_rich_text_options import base_rich_text_options
 from networkapi.wagtailpages.pagemodels.mixin.snippets import LocalizedSnippet
-from networkapi.wagtailpages.utils import get_language_from_request
+from networkapi.wagtailpages.utils import (
+    TitleWidget,
+    insert_panels_after,
+    get_original_by_slug,
+    get_language_from_request,
+)
 
 TRACK_RECORD_CHOICES = [
     ('Great', 'Great'),
@@ -57,64 +47,6 @@ TRACK_RECORD_CHOICES = [
     ('Needs Improvement', 'Needs Improvement'),
     ('Bad', 'Bad')
 ]
-
-
-def get_categories_for_locale(language_code):
-    """
-    Start with the English list of categories, and replace any of them
-    with their localized counterpart, where possible, so that we don't
-    end up with an incomplete category list due to missing locale records.
-    """
-    DEFAULT_LANGUAGE_CODE = settings.LANGUAGE_CODE
-    (DEFAULT_LOCALE, DEFAULT_LOCALE_ID) = get_default_locale()
-
-    default_locale_list = BuyersGuideProductCategory.objects.filter(
-        hidden=False,
-        locale=DEFAULT_LOCALE,
-    )
-
-    if language_code == DEFAULT_LANGUAGE_CODE:
-        return default_locale_list
-
-    try:
-        actual_locale = Locale.objects.get(language_code=language_code)
-    except Locale.DoesNotExist:
-        actual_locale = Locale.objects.get(language_code=DEFAULT_LANGUAGE_CODE)
-
-    return [
-        BuyersGuideProductCategory.objects.filter(
-            translation_key=cat.translation_key,
-            locale=actual_locale,
-        ).first() or cat for cat in default_locale_list
-    ]
-
-
-def get_product_subset(cutoff_date, authenticated, key, products, language_code='en'):
-    """
-    filter a queryset based on our current cutoff date,
-    as well as based on whether a user is authenticated
-    to the system or not (authenticated users get to
-    see all products, including draft products)
-    """
-    try:
-        locale = Locale.objects.get(language_code=language_code)
-    except Locale.DoesNotExist:
-        locale = Locale.objects.get(language_code=settings.LANGUAGE_CODE)
-
-    products = products.filter(review_date__gte=cutoff_date, locale=locale)
-
-    if not authenticated:
-        products = products.live()
-
-    products = sort_average(products)
-    return cache.get_or_set(key, products, 86400)
-
-
-def sort_average(products):
-    """
-    `products` is a QuerySet of ProductPages.
-    """
-    return sorted(products, key=lambda p: p.creepiness)
 
 
 @register_snippet
@@ -944,6 +876,7 @@ class ProductPage(AirtableMixin, FoundationMetadataPageMixin, Page):
         return "unknown"
 
     def get_context(self, request, *args, **kwargs):
+        BuyersGuidePage = apps.get_model(app_label='wagtailpages', model_name='BuyersGuidePage')
         context = super().get_context(request, *args, **kwargs)
         context['product'] = self
         language_code = get_language_from_request(request)
@@ -1361,278 +1294,3 @@ class ExcludedCategories(TranslatableMixin, Orderable):
 
     class Meta(TranslatableMixin.Meta):
         verbose_name = 'Excluded Category'
-
-
-class BuyersGuidePage(RoutablePageMixin, FoundationMetadataPageMixin, Page):
-    """
-    Note: We'll likely be converting the "about" pages to Wagtail Pages.
-    When that happens, we should remove the RoutablePageMixin and @routes
-    """
-
-    template = 'pages/buyersguide/home.html'
-    subpage_types = [
-        'wagtailpages.GeneralProductPage',
-        'wagtailpages.BuyersGuideEditorialContentIndexPage',
-    ]
-
-    cutoff_date = models.DateField(
-        'Product listing cutoff date',
-        help_text='Only show products that were reviewed on, or after this date.',
-        default=datetime(2020, 10, 29),
-    )
-
-    hero_image = models.ForeignKey(
-        'wagtailimages.Image',
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name='pni_hero_image'
-    )
-
-    header = models.CharField(
-        max_length=120,
-        blank=True,
-        help_text='The header text for the PNI homepage',
-    )
-
-    intro_text = models.TextField(
-        max_length=500,
-        blank=True,
-        help_text='A short blurb to show under the header',
-    )
-
-    dark_theme = models.BooleanField(
-        default=False,
-        help_text='Does the intro need to be white text (for dark backgrounds)?'
-    )
-
-    def get_banner(self):
-        return self.hero_image
-
-    content_panels = [
-        FieldPanel('title'),
-        FieldPanel('cutoff_date'),
-        ImageChooserPanel('hero_image'),
-        FieldPanel('header'),
-        FieldPanel('intro_text'),
-        FieldPanel('dark_theme'),
-        MultiFieldPanel(
-            [
-                InlinePanel("excluded_categories", label="Category", min_num=0)
-            ],
-            heading="Excluded Categories"
-        ),
-    ]
-
-    translatable_fields = [
-        TranslatableField('title'),
-        SynchronizedField('cutoff_date'),
-        SynchronizedField('hero_image'),
-        TranslatableField('header'),
-        TranslatableField('intro_text'),
-        SynchronizedField('dark_theme'),
-        SynchronizedField('excluded_categories'),
-        # Promote tab fields
-        TranslatableField('seo_title'),
-        TranslatableField('search_description'),
-        SynchronizedField('search_image'),
-    ]
-
-    @route(r'^about/$', name='how-to-use-view')
-    def about_page(self, request):
-        context = self.get_context(request)
-        context['pagetype'] = 'about'
-        context['pageTitle'] = pgettext(
-            '*privacy not included can be localized.',
-            'How to use *privacy not included')
-        return render(request, "pages/buyersguide/how_to_use.html", context)
-
-    @route(r'^about/why/$', name='about-why-view')
-    def about_why_page(self, request):
-        context = self.get_context(request)
-        context['pagetype'] = 'about'
-        context['pageTitle'] = pgettext(
-            '*privacy not included can be localized.',
-            'Why we made *privacy not included')
-        return render(request, "pages/buyersguide/why_we_made.html", context)
-
-    @route(r'^about/press/$', name='press-view')
-    def about_press_page(self, request):
-        context = self.get_context(request)
-        context['pagetype'] = 'about'
-        context['pageTitle'] = pgettext('Noun, media', 'Press') + ' | ' + pgettext(
-            'This can be localized. This is a reference to the “*batteries not included” mention on toys.',
-            '*privacy not included')
-        return render(request, "pages/buyersguide/press.html", context)
-
-    @route(r'^about/contact/$', name='contact-view')
-    def about_contact_page(self, request):
-        context = self.get_context(request)
-        context['pagetype'] = 'about'
-        context['pageTitle'] = gettext('Contact us') + ' | ' + pgettext(
-            'This can be localized. This is a reference to the “*batteries not included” mention on toys.',
-            '*privacy not included')
-        return render(request, "pages/buyersguide/contact.html", context)
-
-    @route(r'^about/methodology/$', name='methodology-view')
-    def about_methodology_page(self, request):
-        context = self.get_context(request)
-        context['pagetype'] = 'about'
-        context['pageTitle'] = gettext('Our methodology') + ' | ' + pgettext(
-            'This can be localized. This is a reference to the “*batteries not included” mention on toys.',
-            '*privacy not included')
-        return render(request, "pages/buyersguide/methodology.html", context)
-
-    @route(r'^contest/$', name='contest')
-    def about_contest(self, request):
-        context = self.get_context(request)
-        context['pagetype'] = 'contest'
-        context['pageTitle'] = gettext('Contest terms and conditions') + ' | ' + pgettext(
-            'This can be localized. This is a reference to the “*batteries not included” mention on toys.',
-            '*privacy not included')
-        return render(request, "pages/buyersguide/contest.html", context)
-
-    @route(r'^products/(?P<slug>[-\w\d]+)/$', name='product-view')
-    def product_view(self, request, slug):
-        # Find product by it's slug and redirect to the product page
-        # If no product is found, redirect to the BuyersGuide page
-        locale = get_locale_from_request(request)
-        product = get_object_or_404(ProductPage, slug=slug, locale=locale)
-        url = relocalize_url(product.url, locale.language_code)
-        return redirect(url)
-
-    @route(r'^categories/(?P<slug>[\w\W]+)/', name='category-view')
-    def categories_page(self, request, slug):
-        context = self.get_context(request, bypass_products=True)
-        language_code = get_language_from_request(request)
-        locale_id = Locale.objects.get(language_code=language_code).id
-        slug = slugify(slug)
-
-        (DEFAULT_LOCALE, DEFAULT_LOCALE_ID) = get_default_locale()
-
-        # because we may be working with localized content, and the slug
-        # will always be our english slug, we need to find the english
-        # category first, and then find its corresponding localized version
-        try:
-            original_category = BuyersGuideProductCategory.objects.get(slug=slug, locale_id=DEFAULT_LOCALE_ID)
-        except BuyersGuideProductCategory.DoesNotExist:
-            original_category = get_object_or_404(BuyersGuideProductCategory, name__iexact=slug)
-
-        if locale_id != DEFAULT_LOCALE_ID:
-            try:
-                category = BuyersGuideProductCategory.objects.get(
-                    translation_key=original_category.translation_key,
-                    locale_id=DEFAULT_LOCALE_ID,
-                )
-            except BuyersGuideProductCategory.DoesNotExist:
-                category = original_category
-        else:
-            category = original_category
-
-        authenticated = request.user.is_authenticated
-        key = f'cat_product_dicts_{slug}_auth' if authenticated else f'cat_product_dicts_{slug}_live'
-        key = f'{language_code}_{key}'
-        products = cache.get(key)
-        exclude_cat_ids = [excats.category.id for excats in self.excluded_categories.all()]
-
-        if products is None:
-            products = get_product_subset(
-                self.cutoff_date,
-                authenticated,
-                key,
-                ProductPage.objects.exclude(product_categories__category__id__in=exclude_cat_ids),
-                language_code=language_code
-            )
-
-        context['category'] = slug
-        context['current_category'] = category
-        context['products'] = products
-        context['pageTitle'] = f'{category.localized.name} | {gettext("Privacy & security guide")}'\
-                               f' | Mozilla Foundation'
-        context['template_cache_key_fragment'] = f'{category.slug}_{request.LANGUAGE_CODE}'
-
-        # Checking if category has custom metadata, if so, update the share image and description.
-        if category.share_image:
-            setattr(self, 'search_image_id', category.localized.share_image_id)
-        if category.description:
-            setattr(self, 'search_description', category.localized.description)
-
-        return render(request, "pages/buyersguide/category_page.html", context)
-
-    def get_sitemap_urls(self, request):
-        """
-        Add categories and route views to the sitemap for better search indexing.
-        """
-        sitemap = super().get_sitemap_urls(request)
-        last_modified = self.last_published_at or self.latest_revision_created_at
-        # Add all the available Buyers Guide categories to the sitemap
-        categories = BuyersGuideProductCategory.objects.filter(hidden=False)
-        for category in categories:
-            sitemap.append(
-                {
-                    "location": self.full_url + self.reverse_subpage("category-view", args=(category.slug,)),
-                    "lastmod": last_modified,
-                }
-            )
-
-        # Add all the available "subpages" (routes) to the sitemap, excluding categories and products.
-        # Categories are added above. Products will be their own Wagtail pages and get their own URLs because of that.
-        about_page_views = [
-            'how-to-use-view', 'about-why-view', 'press-view',
-            'contact-view', 'methodology-view'
-        ]
-        for about_page in about_page_views:
-            sitemap.append(
-                {
-                    "location": self.full_url + self.reverse_subpage(about_page),
-                    "lastmod": last_modified,
-                }
-            )
-        return sitemap
-
-    def get_context(self, request, *args, **kwargs):
-        context = super().get_context(request, *args, **kwargs)
-        language_code = get_language_from_request(request)
-
-        authenticated = request.user.is_authenticated
-        key = 'home_product_dicts_authed' if authenticated else 'home_product_dicts_live'
-        key = f'{key}_{language_code}'
-        products = cache.get(key)
-        exclude_cat_ids = [excats.category.id for excats in self.excluded_categories.all()]
-
-        if not kwargs.get('bypass_products', False) and products is None:
-            products = get_product_subset(
-                self.cutoff_date,
-                authenticated,
-                key,
-                ProductPage.objects.exclude(product_categories__category__id__in=exclude_cat_ids),
-                language_code=language_code
-            )
-
-        context['categories'] = get_categories_for_locale(language_code)
-        context['current_category'] = None
-        context['products'] = products
-        context['web_monetization_pointer'] = settings.WEB_MONETIZATION_POINTER
-        pni_home_page = BuyersGuidePage.objects.first()
-        context['about_page'] = pni_home_page
-        context['home_page'] = pni_home_page
-        context['template_cache_key_fragment'] = f'pni_home_{request.LANGUAGE_CODE}'
-        return context
-
-    def get_editorial_content_index(self):
-        BuyersGuideEditorialContentIndexPage = apps.get_model(
-            app_label='wagtailpages',
-            model_name='BuyersGuideEditorialContentIndexPage',
-        )
-        indexes = BuyersGuideEditorialContentIndexPage.objects.descendant_of(self)
-        return indexes.first()
-
-    class Meta:
-        verbose_name = "Buyers Guide Page"
-
-
-def get_pni_home_page():
-    """
-    Used in AIRTABLE settings for nesting child pages under a new parent page.
-    """
-    return BuyersGuidePage.objects.first().id
